chore: tighten test helpers and propagate ctx into packages.Load
- internal/rule/context.go: align AnalysisContext.Fset godoc with the
  defensive nil branch in Position so the doc and code stop disagreeing.
- internal/lints/repeated_regex_compilation.go: thread ctx into
  packages.Config.Context so packages.Load respects -timeout and
  cancellation, matching the exec.CommandContext propagation already
  in golangci-lint.
- internal/lints/default_golangci.go + parse_helper_test.go: move the
  test-only ParseFile helper out of production code into a *_test.go
  file. The five callers in *_test.go keep working; the shipped
  binary no longer carries the helper or its go/parser dependency.
- internal/engine_test.go, internal/engine_filename_test.go: drop the
  bespoke createTempDir(t, prefix) wrapper and use the standard
  t.TempDir() / b.TempDir() at all nine call sites.

No behavior change in the engine hot path. BenchmarkRun unchanged.
